Lambda/Functions/cmd_sym/requirements.txt
appdirs==1.4.4
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
cssselect==1.1.0
cycler==0.11.0
fake-useragent==0.1.11
feedparser==6.0.8
fonttools==4.31.2
idna==3.3
importlib-metadata==4.11.3
kiwisolver==1.4.1
lxml==4.8.0
matplotlib==3.5.1
multitasking==0.0.10
numpy==1.22.3
packaging==21.3
pandas==1.4.1
parse==1.19.0
Pillow==9.0.1
prettytable==3.2.0
pyee==8.2.2
pyparsing==3.0.7
pyppeteer==1.0.2
pyquery==1.4.3
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
requests-futures==1.0.0
requests-html==0.10.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.1
tqdm==4.63.1
tzdata==2022.1
urllib3==1.26.9
w3lib==1.22.0
wcwidth==0.2.5
websockets==10.2
yahoo-fin==0.8.9.1
yahooquery==2.2.15
yfinance==0.1.70
zipp==3.7.0
